+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.0.9
 
 - Add the text "Next" before the Next Tetromino Preview area
+- "Show Next Block Preview" menu item under "View" menu
 - Fix Tetris Help menu (was called Options by mistake)
 
 ## 0.0.8